+++ qiita_ware/util.py
@@ -11,11 +11,8 @@ from __future__ import division
 import sys
 from collections import defaultdict
 from heapq import heappush, heappop
-from contextlib import contextmanager
 
-import h5py
 import numpy as np
-from skbio.io.util import _is_string_or_bytes
 from future.utils import viewitems
 
 
@@ -94,61 +91,3 @@ def per_sample_sequences(iter_, max_seqs, min_seqs=1, random_buf_size=100000):
 
         for _, sequence_id, sequence in heap:
             yield (sequence_id, sequence)
-
-
-def _get_filehandle(filepath_or, *args, **kwargs):
-    """Open file if `filepath_or` looks like a string/unicode/bytes, else
-    pass through.
-    """
-    if _is_string_or_bytes(filepath_or):
-        if h5py.is_hdf5(filepath_or):
-            fh, own_fh = h5py.File(filepath_or, *args, **kwargs), True
-        else:
-            fh, own_fh = open(filepath_or, *args, **kwargs), True
-    else:
-        fh, own_fh = filepath_or, False
-    return fh, own_fh
-
-
-@contextmanager
-def open_file(filepath_or, *args, **kwargs):
-    """Context manager, like ``open``, but lets file handles and file like
-    objects pass untouched.
-
-    It is useful when implementing a function that can accept both
-    strings and file-like objects (like numpy.loadtxt, etc).
-
-    This method differs slightly from scikit-bio's implementation in that it
-    handles HDF5 files appropriately.
-
-    Parameters
-    ----------
-    filepath_or : str/bytes/unicode string or file-like
-         If string, file to be opened using ``h5py.File`` if the file is an
-         HDF5 file, otherwise builtin ``open`` will be used. If it is not a
-         string, the object is just returned untouched.
-
-    Other parameters
-    ----------------
-    args, kwargs : tuple, dict
-        When `filepath_or` is a string, any extra arguments are passed
-        on to the ``open`` builtin.
-
-    Examples
-    --------
-    >>> with open_file('filename') as f:  # doctest: +SKIP
-    ...     pass
-    >>> fh = open('filename')             # doctest: +SKIP
-    >>> with open_file(fh) as f:          # doctest: +SKIP
-    ...     pass
-    >>> fh.closed                         # doctest: +SKIP
-    False
-    >>> fh.close()                        # doctest: +SKIP
-
-    """
-    fh, own_fh = _get_filehandle(filepath_or, *args, **kwargs)
-    try:
-        yield fh
-    finally:
-        if own_fh:
-            fh.close()
