+++ setup.py
@@ -106,9 +106,9 @@ setup(name='qiita-spots',
       install_requires=['psycopg2 < 2.7', 'click >= 3.3', 'future',
                         'bcrypt', 'pandas >= 0.17', 'numpy >= 1.7',
                         'tornado==3.1.1', 'toredis', 'redis', 'six',
-                        'ipython[all] >= 2.4.1, < 2.5', 'pyparsing',
+                        'ipython[all] == 3.2.0', 'pyparsing',
                         'h5py >= 2.3.1', 'biom-format', 'natsort', 'networkx',
-                        'scikit-bio >= 0.2.3, < 0.3.0', 'wtforms == 2.0.1',
+                        'scikit-bio == 0.4.2', 'wtforms == 2.0.1',
                         'moi', 'sphinx-bootstrap-theme', 'Sphinx >= 1.2.2',
                         'gitpython', 'qiita-files'],
       dependency_links=[
